# clean up some unsafety, help LLVM elide bounds checks